# internal/web/users/api.gen.go
// Package users provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package users

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	strictecho "github.com/oapi-codegen/runtime/strictmiddleware/echo"
)

// Defines values for ErrorResponseErrorCode.
const (
	NOCANDIDATE ErrorResponseErrorCode = "NO_CANDIDATE"
	NOTASSIGNED ErrorResponseErrorCode = "NOT_ASSIGNED"
	NOTFOUND    ErrorResponseErrorCode = "NOT_FOUND"
	PREXISTS    ErrorResponseErrorCode = "PR_EXISTS"
	PRMERGED    ErrorResponseErrorCode = "PR_MERGED"
	TEAMEXISTS  ErrorResponseErrorCode = "TEAM_EXISTS"
)

// Defines values for PullRequestShortStatus.
const (
	MERGED PullRequestShortStatus = "MERGED"
	OPEN   PullRequestShortStatus = "OPEN"
)

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error struct {
		Code    ErrorResponseErrorCode `json:"code"`
		Message string                 `json:"message"`
	} `json:"error"`
}

// ErrorResponseErrorCode defines model for ErrorResponse.Error.Code.
type ErrorResponseErrorCode string

// PullRequestShort defines model for PullRequestShort.
type PullRequestShort struct {
	AuthorId        string                 `json:"author_id"`
	PullRequestId   string                 `json:"pull_request_id"`
	PullRequestName string                 `json:"pull_request_name"`
	Status          PullRequestShortStatus `json:"status"`
}

// PullRequestShortStatus defines model for PullRequestShort.Status.
type PullRequestShortStatus string

// User defines model for User.
type User struct {
	IsActive bool   `json:"is_active"`
	TeamName string `json:"team_name"`
	UserId   string `json:"user_id"`
	Username string `json:"username"`
}

// UserIdQuery defines model for UserIdQuery.
type UserIdQuery = string

// PostUsersDeactivateJSONBody defines parameters for PostUsersDeactivate.
type PostUsersDeactivateJSONBody struct {
	TeamName string   `json:"team_name"`
	UserIds  []string `json:"user_ids"`
}

// GetUsersGetReviewParams defines parameters for GetUsersGetReview.
type GetUsersGetReviewParams struct {
	// UserId Идентификатор пользователя
	UserId UserIdQuery `form:"user_id" json:"user_id"`
}

// PostUsersSetIsActiveJSONBody defines parameters for PostUsersSetIsActive.
type PostUsersSetIsActiveJSONBody struct {
	IsActive bool   `json:"is_active"`
	UserId   string `json:"user_id"`
}

// PostUsersDeactivateJSONRequestBody defines body for PostUsersDeactivate for application/json ContentType.
type PostUsersDeactivateJSONRequestBody PostUsersDeactivateJSONBody

// PostUsersSetIsActiveJSONRequestBody defines body for PostUsersSetIsActive for application/json ContentType.
type PostUsersSetIsActiveJSONRequestBody PostUsersSetIsActiveJSONBody

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Массово деактивировать пользователей команды и переназначить открытые PR
	// (POST /users/deactivate)
	PostUsersDeactivate(ctx echo.Context) error
	// Получить PR'ы, где пользователь назначен ревьювером
	// (GET /users/getReview)
	GetUsersGetReview(ctx echo.Context, params GetUsersGetReviewParams) error
	// Установить флаг активности пользователя
	// (POST /users/setIsActive)
	PostUsersSetIsActive(ctx echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// PostUsersDeactivate converts echo context to params.
func (w *ServerInterfaceWrapper) PostUsersDeactivate(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostUsersDeactivate(ctx)
	return err
}

// GetUsersGetReview converts echo context to params.
func (w *ServerInterfaceWrapper) GetUsersGetReview(ctx echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetUsersGetReviewParams
	// ------------- Required query parameter "user_id" -------------

	err = runtime.BindQueryParameter("form", true, true, "user_id", ctx.QueryParams(), &params.UserId)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter user_id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetUsersGetReview(ctx, params)
	return err
}

// PostUsersSetIsActive converts echo context to params.
func (w *ServerInterfaceWrapper) PostUsersSetIsActive(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostUsersSetIsActive(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/users/deactivate", wrapper.PostUsersDeactivate)
	router.GET(baseURL+"/users/getReview", wrapper.GetUsersGetReview)
	router.POST(baseURL+"/users/setIsActive", wrapper.PostUsersSetIsActive)

}

type PostUsersDeactivateRequestObject struct {
	Body *PostUsersDeactivateJSONRequestBody
}

type PostUsersDeactivateResponseObject interface {
	VisitPostUsersDeactivateResponse(w http.ResponseWriter) error
}

type PostUsersDeactivate200JSONResponse struct {
	AffectedPrCount          int    `json:"affected_pr_count"`
	DeactivatedCount         int    `json:"deactivated_count"`
	ReassignedReviewersCount int    `json:"reassigned_reviewers_count"`
	TeamName                 string `json:"team_name"`
}

func (response PostUsersDeactivate200JSONResponse) VisitPostUsersDeactivateResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type PostUsersDeactivate400JSONResponse ErrorResponse

func (response PostUsersDeactivate400JSONResponse) VisitPostUsersDeactivateResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type PostUsersDeactivate404JSONResponse ErrorResponse

func (response PostUsersDeactivate404JSONResponse) VisitPostUsersDeactivateResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type GetUsersGetReviewRequestObject struct {
	Params GetUsersGetReviewParams
}

type GetUsersGetReviewResponseObject interface {
	VisitGetUsersGetReviewResponse(w http.ResponseWriter) error
}

type GetUsersGetReview200JSONResponse struct {
	PullRequests []PullRequestShort `json:"pull_requests"`
	UserId       string             `json:"user_id"`
}

func (response GetUsersGetReview200JSONResponse) VisitGetUsersGetReviewResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type PostUsersSetIsActiveRequestObject struct {
	Body *PostUsersSetIsActiveJSONRequestBody
}

type PostUsersSetIsActiveResponseObject interface {
	VisitPostUsersSetIsActiveResponse(w http.ResponseWriter) error
}

type PostUsersSetIsActive200JSONResponse struct {
	User *User `json:"user,omitempty"`
}

func (response PostUsersSetIsActive200JSONResponse) VisitPostUsersSetIsActiveResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type PostUsersSetIsActive404JSONResponse ErrorResponse

func (response PostUsersSetIsActive404JSONResponse) VisitPostUsersSetIsActiveResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Массово деактивировать пользователей команды и переназначить открытые PR
	// (POST /users/deactivate)
	PostUsersDeactivate(ctx context.Context, request PostUsersDeactivateRequestObject) (PostUsersDeactivateResponseObject, error)
	// Получить PR'ы, где пользователь назначен ревьювером
	// (GET /users/getReview)
	GetUsersGetReview(ctx context.Context, request GetUsersGetReviewRequestObject) (GetUsersGetReviewResponseObject, error)
	// Установить флаг активности пользователя
	// (POST /users/setIsActive)
	PostUsersSetIsActive(ctx context.Context, request PostUsersSetIsActiveRequestObject) (PostUsersSetIsActiveResponseObject, error)
}

type StrictHandlerFunc = strictecho.StrictEchoHandlerFunc
type StrictMiddlewareFunc = strictecho.StrictEchoMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// PostUsersDeactivate operation middleware
func (sh *strictHandler) PostUsersDeactivate(ctx echo.Context) error {
	var request PostUsersDeactivateRequestObject

	var body PostUsersDeactivateJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostUsersDeactivate(ctx.Request().Context(), request.(PostUsersDeactivateRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostUsersDeactivate")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(PostUsersDeactivateResponseObject); ok {
		return validResponse.VisitPostUsersDeactivateResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// GetUsersGetReview operation middleware
func (sh *strictHandler) GetUsersGetReview(ctx echo.Context, params GetUsersGetReviewParams) error {
	var request GetUsersGetReviewRequestObject

	request.Params = params

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetUsersGetReview(ctx.Request().Context(), request.(GetUsersGetReviewRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetUsersGetReview")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetUsersGetReviewResponseObject); ok {
		return validResponse.VisitGetUsersGetReviewResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// PostUsersSetIsActive operation middleware
func (sh *strictHandler) PostUsersSetIsActive(ctx echo.Context) error {
	var request PostUsersSetIsActiveRequestObject

	var body PostUsersSetIsActiveJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostUsersSetIsActive(ctx.Request().Context(), request.(PostUsersSetIsActiveRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostUsersSetIsActive")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(PostUsersSetIsActiveResponseObject); ok {
		return validResponse.VisitPostUsersSetIsActiveResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}